Rename comparision -> comparison
