+++ core/src/comparison_index.rs
